requirements.txt
astroid==2.0.4
brotlipy==0.7.0
certifi==2018.8.24
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.10
colorama==0.4.4
conda==4.3.16
cryptography==3.4.7
defusedxml==0.5.0
Django==2.2.13
django-allauth==0.47.0
django-cors-headers==3.10.1
django-cors-middleware==1.5.0
django-rest-auth==0.9.5
django-session-timeout==0.1.0
djangorestframework==3.9.1
gcloud==0.18.3
googleapis-common-protos==1.54.0
gunicorn==19.7.1
httplib2==0.20.2
idna==2.7
isort==4.3.4
jws==0.1.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauth2client==4.1.3
oauthlib==2.1.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.20
pycryptodome==3.12.0
PyJWT==2.3.0
pyOpenSSL==20.0.1
pyparsing==3.0.6
Pyrebase4==4.5.0
PySocks==1.7.1
python-dotenv==0.19.2
python-jwt==2.0.1
python3-openid==3.1.0
pytz==2018.5
requests==2.27.1
requests-oauthlib==1.0.0
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
sqlparse==0.3.0
tqdm==4.61.2
urllib3==1.24.2
whitenoise==3.3.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.10.11
